assignment_6

+++ common/core/vec3.cpp
@@ -3,7 +3,7 @@
 // methods don't modify the object itself
 
 vec3::vec3() : x(0), y(0), z(0) {}
-vec3::vec3(double x, double y, double z) : x(x), y(y), z(z) {}
+vec3::vec3(float x, float y, float z) : x(x), y(y), z(z) {}
 
 vec3 vec3::operator+(const vec3& v) const {
     return vec3(x + v.x, y + v.y, z + v.z);
@@ -13,11 +13,11 @@ vec3 vec3::operator-(const vec3& v) const {
     return vec3(x - v.x, y - v.y, z - v.z);
 }
 
-vec3 vec3::operator*(double t) const {
+vec3 vec3::operator*(float t) const {
     return vec3(x * t, y * t, z * t);
 }
 
-vec3 vec3::operator/(double t) const {
+vec3 vec3::operator/(float t) const {
     return vec3(x / t, y / t, z / t);
 }
 
@@ -26,7 +26,7 @@ std::ostream& operator<<(std::ostream& os, const vec3& v)  {
     return os;
 }
 
-double vec3::dot(const vec3& v) const {
+float vec3::dot(const vec3& v) const {
     return x * v.x + y * v.y + z * v.z;
 }
 
@@ -38,7 +38,7 @@ vec3 vec3::cross(const vec3& v) const {
     );
 }
 
-double vec3::length() const {
+float vec3::length() const {
     return std::sqrt(dot(*this));
 }
